sim/requirements.txt
cupy-cuda115==10.0.0
fastrlock==0.8
llvmlite==0.37.0
numba==0.54.1
numpy==1.20.3
